requirements/dev.txt
-r ../requirements.txt

python-slugify==1.2.0
GitPython==1.0.2
Sphinx==1.3.5
ipdb==0.8.1
